requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.8
django-filter==25.2
django-phonenumber-field==8.3.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
phonenumbers==9.0.18
phonenumberslite==9.0.20
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
pypng==0.20220715.0
python-dotenv==1.2.1
qrcode==8.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
